chore(WorldLoadListener): Rename WorldManager.java to WorldLoadListener.java
